Fix some MSVC12 conversion warnings

+++ library/ecdsa.c
@@ -49,7 +49,7 @@
  *
  * Argument is the minimum size in bytes of the MD output.
  */
-static const md_info_t *md_info_by_size( int min_size )
+static const md_info_t *md_info_by_size( size_t min_size )
 {
     const md_info_t *md_cur, *md_picked = NULL;
     const int *md_alg;
@@ -57,7 +57,7 @@ static const md_info_t *md_info_by_size( int min_size )
     for( md_alg = md_list(); *md_alg != 0; md_alg++ )
     {
         if( ( md_cur = md_info_from_type( *md_alg ) ) == NULL ||
-            md_cur->size < min_size ||
+            (size_t) md_cur->size < min_size ||
             ( md_picked != NULL && md_cur->size > md_picked->size ) )
             continue;
